+++ app/signin/page.tsx
@@ -1,5 +1,5 @@
-import AuthPageLayout from "@features/auth/AuthPageLayout";
-import SigninForm from "@features/auth/SigninForm";
+import AuthPageLayout from "@features/auth/components/AuthPageLayout";
+import SigninForm from "@features/auth/components/SigninForm";
 
 export default function Signin() {
   return (
